+++ pyproject.toml
@@ -5,6 +5,7 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
+    "aiohttp>=3.12.14",
     "async-lru>=2.0.5",
     "mcp[cli]>=1.9.1",
 ]
